+++ android/app/src/main/java/org/smartregister/fhircore/activity/PatientListActivity.kt
@@ -16,6 +16,7 @@
 
 package org.smartregister.fhircore.activity
 
+import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import android.view.View
@@ -30,6 +31,7 @@ import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import androidx.viewpager2.widget.ViewPager2
 import org.smartregister.fhircore.R
+import org.smartregister.fhircore.fragment.PatientDetailFragment
 import org.smartregister.fhircore.fragment.PatientListFragment
 import org.smartregister.fhircore.util.Utils
 import org.smartregister.fhircore.util.Utils.addOnDrawableClickedListener
@@ -85,12 +87,19 @@ class PatientListActivity : BaseSimpleActivity() {
   private fun addPatient(view: View) {
     // TO DO: Open patient registration form
     val context = view.context
+    startRegistrationActivity(context, null)
+  }
+
+  fun startRegistrationActivity(context: Context, preAssignedId: String?) {
     context.startActivity(
       Intent(context, QuestionnaireActivity::class.java).apply {
         putExtra(
           QuestionnaireActivity.QUESTIONNAIRE_TITLE_KEY,
           this@PatientListActivity.getString(R.string.client_info)
         )
+
+        if (!preAssignedId.isNullOrEmpty())
+          putExtra(PatientDetailFragment.ARG_PRE_ASSIGNED_ID, preAssignedId)
         putExtra(QuestionnaireActivity.QUESTIONNAIRE_FILE_PATH_KEY, "patient-registration.json")
       }
     )